requirements.txt
# 웹 프레임워크
fastapi
uvicorn[standard]

# Data Processing
numpy
pandas
scikit-learn

# AI & LLM
openai
google-generativeai
python-dotenv

# Database & Vector Search
qdrant-client
FlagEmbedding
transformers>=4.44.2,<5
torch
psycopg[binary]

# Utility
pydantic-settings
httpx
aio-pika
prometheus-client
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp-proto-grpc
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-logging


# pip install -r requirements.txt
# 설치된 모든 라이브러리의 정확한 버전을 확인하고 싶을 때
# pip freeze > requirements_lock.txt
